Lymph (cannon) Idle animation
This should be mounted on the walls (behind the black) so it sticks out
to a little more than a half circle. The white blood cells shoot from
the center (it would be cool if they grew from the inside). Another
thing is it would be cool for the animation to speed when it is ready to
shoot. This would be a good indicator for the player that they will need
to take evasive action.